+++ echopype/mask/__init__.py
@@ -1,4 +1,3 @@
-from .api import apply_mask, frequency_differencing
-from .mask_seabed import get_seabed_mask
+from .api import apply_mask, frequency_differencing, get_seabed_mask
 
 __all__ = ["frequency_differencing", "apply_mask", "get_seabed_mask"]
